+++ wain-syntax-text/src/ast.rs
@@ -201,7 +201,7 @@ pub struct Instruction<'a> {
 // https://webassembly.github.io/spec/core/text/instructions.html#text-memarg
 #[cfg_attr(test, derive(Debug))]
 pub struct Mem {
-    pub align: Option<u32>,
+    pub align: Option<u8>,
     pub offset: Option<u32>,
 }
